Create Musical
"IYANU OLUWA HOT the fiery hip-hop/rap artist, is making waves in the music scene with his unapologetic lyrics and infectious beats. With a passion for storytelling, IYANU OLUWA HOT delivers raw, honest verses that resonate with listeners from all walks of life. His music is a reflection of his experiences, emotions, and perspectives, making his art relatable and impactful. With a string of released songs and a growing fan base, IYANU OLUWA HOT is solidifying his position as a rising star in the hip-hop world. Get ready to feel the heat as IYANU OLUWA HOT continues to drop hot bars and push the boundaries of what's possible in hip-hop.